multicluster kind e2e (#22)
* mc test
* nicely skippable
* cleanup
* brackets
* move helper struct
* use a ref that should not be committed
* local util, cleanup
* retry action
* deps
* pin tools
* cleaner boiler
* less magic'
* testing readme for render suite
* keepgoing
