+++ test/test_config.go
@@ -2,23 +2,26 @@ package test
 
 import (
 	"context"
+	"os"
 
 	"github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+	"k8s.io/client-go/tools/clientcmd/api"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 )
 
-func MustConfig() *rest.Config {
-	cfg, err := config.GetConfig()
+func MustConfig(context string) *rest.Config {
+	cfg, err := config.GetConfigWithContext(context)
 	Expect(err).NotTo(HaveOccurred())
 	return cfg
 }
 
 func MustManager(ns string) (manager.Manager, func()) {
-	cfg := MustConfig()
+	cfg := MustConfig("")
 	return ManagerWithOpts(cfg, manager.Options{
 		Namespace: ns,
 		// Disable metrics and health probe to allow tests to run in parallel.
@@ -45,7 +48,7 @@ func ManagerWithOpts(cfg *rest.Config, opts manager.Options) (manager.Manager, f
 }
 
 func MustManagerNotStarted(ns string) manager.Manager {
-	mgr, err := manager.New(MustConfig(), manager.Options{
+	mgr, err := manager.New(MustConfig(""), manager.Options{
 		Namespace: ns,
 		// Disable metrics and health probe to allow tests to run in parallel.
 		MetricsBindAddress:     "0",
@@ -53,4 +56,13 @@ func MustManagerNotStarted(ns string) manager.Manager {
 	})
 	Expect(err).NotTo(HaveOccurred())
 	return mgr
+}
+
+func MustClientConfigWithContext(context string) *api.Config {
+	loadingRules := clientcmd.NewDefaultClientConfigLoadingRules()
+	loadingRules.ExplicitPath = os.Getenv("KUBECONFIG")
+	configOverrides := &clientcmd.ConfigOverrides{CurrentContext: context}
+	cfg, err := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(loadingRules, configOverrides).ConfigAccess().GetStartingConfig()
+	Expect(err).NotTo(HaveOccurred())
+	return cfg
 }
